Change file structure
Change file structure to properly display the composition of the application. I decided to keep the main pages at the highest level, even though the app component basically wraps them. This is so the app component doesn't just become another src folder, and it's more obvious what pages are available in the app